+++ web/documentserver-example/csharp-mvc/Models/FileUtility.cs
@@ -239,6 +239,13 @@ namespace OnlineEditorsExampleMVC.Models
             return cachedFormats;
         }
 
+        public static List<string> GetFormatActions(string extension)
+        {
+            return All()
+                .Find(format => format.Extension() == extension)
+                .Actions;
+        }
+
         private static string GetPath()
         {
             string path = Path.Combine(GetDirectory(), "onlyoffice-docs-formats.json");