+++ web/documentserver-example/go/server/managers/config.go
@@ -73,10 +73,7 @@ type JwtManager interface {
 }
 
 type ConversionManager interface {
-	GetFileType(filename string) string
 	GetInternalExtension(fileType string) string
-	IsCanConvert(ext string) bool
-	IsCanFill(ext string) bool
 	GetConverterUri(
 		docUri, fromExt, toExt, docKey string,
 		isAsync bool, title string, filePass string,
@@ -87,6 +84,15 @@ type CommandManager interface {
 	CommandRequest(method string, docKey string, meta interface{}) (*http.Response, error)
 }
 
+type FormatManager interface {
+	GetAllFormats() []models.Format
+	GetFormat(ext string) models.Format
+	GetActions(ext string) []string
+	GetTypeByExtension(ext string) string
+	GetFileType(filename string) string
+	HasAction(ext string, action string) bool
+}
+
 type Managers struct {
 	DocumentManager
 	HistoryManager
@@ -95,12 +101,13 @@ type Managers struct {
 	JwtManager
 	ConversionManager
 	CommandManager
+	FormatManager
 }
 
 func New(umanager UserManager, smanager StorageManager,
 	hmanager HistoryManager, dmanager DocumentManager,
 	jmanager JwtManager, cmanager ConversionManager,
-	commanager CommandManager) *Managers {
+	commanager CommandManager, fmanager FormatManager) *Managers {
 	return &Managers{
 		HistoryManager:    hmanager,
 		StorageManager:    smanager,
@@ -109,5 +116,6 @@ func New(umanager UserManager, smanager StorageManager,
 		JwtManager:        jmanager,
 		ConversionManager: cmanager,
 		CommandManager:    commanager,
+		FormatManager:     fmanager,
 	}
 }