+++ web/documentserver-example/ruby/app/models/file_model.rb
@@ -362,14 +362,22 @@ class FileModel
     # direct url to the image
     insert_image = if enable_direct_url? == true
                      {
-                       fileType: 'svg', # image file type
-                       url: "#{DocumentHelper.get_server_url(true)}/assets/logo.svg", # server url to the image
-                       directUrl: "#{DocumentHelper.get_server_url(false)}/assets/logo.svg" # direct url to the image
+                       images: [
+                         {
+                           fileType: 'svg', # image file type
+                           url: "#{DocumentHelper.get_server_url(true)}/assets/logo.svg", # server url to the image
+                           directUrl: "#{DocumentHelper.get_server_url(false)}/assets/logo.svg" # directUrl to the image
+                         }
+                       ]
                      }
                    else
                      {
-                       fileType: 'svg', # image file type
-                       url: "#{DocumentHelper.get_server_url(true)}/assets/logo.svg" # server url to the image
+                       images: [
+                         {
+                           fileType: 'svg', # image file type
+                           url: "#{DocumentHelper.get_server_url(true)}/assets/logo.svg" # server url to the image
+                         }
+                       ]
                      }
                    end
 
@@ -378,7 +386,7 @@ class FileModel
       insert_image['token'] = JwtHelper.encode(insert_image)
     end
 
-    insert_image.to_json.tr('{', '').tr('}', '')
+    insert_image
   end
 
   # get compared file information